moved direcories_list logic to +page@config.svelte
