+++ teller_desktop/src-tauri/Cargo.toml
@@ -17,12 +17,16 @@ tauri-build = { version = "1.2.1", features = [] }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0.183", features = ["derive"] }
+
 tauri = { version = "1.3", features = [ "shell-all", "window-all", "dialog-all", "http-all", "fs-all", "devtools"] }
 tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+tauri-plugin-single-instance = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+
 log = "0.4.19"
 base64 = "0.21.2"
 config = "0.13.3"
 directories = "5.0.1"
+
 uuid = { version = "1.4.1", features = ["v4", "serde"] }
 tokio = { version = "1.30.0", features = ["full"] }
 chrono = { version = "0.4.26", features = ["serde"] }
