+++ teller_desktop/src-tauri/src/main.rs
@@ -3,9 +3,17 @@
     windows_subsystem = "windows"
 )]
 
+use log::info;
 use std::env;
+use tauri::Manager;
 use tauri_plugin_log::LogTarget;
 use teller::handlers::config::get_config_folder;
+
+#[derive(Clone, serde::Serialize)]
+struct Payload {
+    args: Vec<String>,
+    cwd: String,
+}
 
 fn main() {
     tauri::Builder::default()
@@ -18,6 +26,11 @@ fn main() {
                 ])
                 .build(),
         )
+        .plugin(tauri_plugin_single_instance::init(|app, argv, cwd| {
+            info!("{}, {argv:?}, {cwd}", app.package_info().name);
+            app.emit_all("single-instance", Payload { args: argv, cwd })
+                .unwrap();
+        }))
         .plugin(teller_desktop::config::init())
         .plugin(teller_desktop::backend::folder_handler::init())
         .plugin(teller_desktop::backend::world_handler::init())
