+++ teller_desktop/src/lib/stores.ts
@@ -1,5 +1,5 @@
-import type { WorldItem, WorldLevelData } from '$lib/utils';
-import { writable, type Writable } from 'svelte/store';
+import type { DirectorySettings, WorldItem, WorldLevelData } from '$lib/utils';
+import { writable } from 'svelte/store';
 
 interface Directory {
 	[name: string]: string | string[] | null;
@@ -14,6 +14,8 @@ export let directories = writable<Directory>({});
 export let localDirs = writable<Directory>({});
 
 export let currentDir = writable<CurrentDir>({ category: 'default', path: 'default' });
+
+export let directorySettings = writable<DirectorySettings>({ categories: {} });
 
 export interface WorldCacheItem {
 	name: string;
